api/requirements-sqlite.txt
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.0
python-jose==3.3.0
python-multipart==0.0.7
passlib==1.7.4
bcrypt==4.0.1
python-dateutil==2.8.2
fastapi-users[sqlalchemy]==12.1.3
fastapi-users[oauth]==12.1.3
httpx-oauth>=0.13
setuptools

# Email service dependencies
boto3==1.34.0  # AWS SES
azure-communication-email==1.0.0  # Azure Email Services
requests==2.31.0  # For Mailgun HTTP API
jinja2==3.1.2  # Email templating
reportlab==4.0.9  # PDF generation for email attachments
